+++ Siv3D/include/Siv3D.hpp
@@ -97,6 +97,7 @@
 # include <Siv3D/FormatInt.hpp>
 # include <Siv3D/FormatFloat.hpp>
 # include <Siv3D/FormatData.hpp>
+# include <Siv3D/Formatter.hpp>
 # include <Siv3D/Format.hpp>
