+++ Siv3D/src/Siv3D/Formatter/SivFormatter.cpp
@@ -9,7 +9,7 @@
 //
 //-----------------------------------------------
 
-# include <Siv3D/Format.hpp>
+# include <Siv3D/Formatter.hpp>
 # include <Siv3D/IntFormatter.hpp>
 # include <Siv3D/FormatInt.hpp>
 # include <Siv3D/FormatFloat.hpp>
