+++ WindowsDesktop/Main.cpp
@@ -7,6 +7,7 @@ SIV3D_OPTION_HEADLESS_MODE(true);
 SIV3D_OPTION_OUTPUT_STDERR(true);
 SIV3D_OPTION_RUN_TEST(false);
 
+
 void Main()
 {
 	::Sleep(4000);
@@ -19,6 +20,10 @@ void Main()
 	//Logger << Time::GetNanosec();
 	//Logger << Time::GetNanosec();
 	//Logger << Time::GetNanosec();
+
+	//auto s = Format(U"aaa", 20, U"aaa", std::wstring{});
+
+	//Apply(10, 20);
 
 	//int n;
 	//std::cin >> n;
