+++ WindowsDesktop/Siv3D.vcxproj.filters
@@ -64,9 +64,6 @@
     <Filter Include="src\Siv3D\StringView">
       <UniqueIdentifier>{ad63d2d2-5695-4e93-81fc-f9f1f0672424}</UniqueIdentifier>
     </Filter>
-    <Filter Include="src\Siv3D\Format">
-      <UniqueIdentifier>{f7fdfc9c-ecac-403a-b8b2-7234a2cfecb7}</UniqueIdentifier>
-    </Filter>
     <Filter Include="src\Siv3D\IntFormatter">
       <UniqueIdentifier>{b4722747-7e40-41a3-a567-89aa12ca64ee}</UniqueIdentifier>
     </Filter>
@@ -111,6 +108,9 @@
     </Filter>
     <Filter Include="src\Siv3D-Platform\WindowsDesktop\DLL">
       <UniqueIdentifier>{e2ef10a8-201e-40ed-9caf-204bbe8d3b7f}</UniqueIdentifier>
+    </Filter>
+    <Filter Include="src\Siv3D\Formatter">
+      <UniqueIdentifier>{f7fdfc9c-ecac-403a-b8b2-7234a2cfecb7}</UniqueIdentifier>
     </Filter>
   </ItemGroup>
   <ItemGroup>
@@ -342,6 +342,9 @@
     <ClInclude Include="..\Siv3D\include\Siv3D\DLL.hpp">
       <Filter>include\Siv3D</Filter>
     </ClInclude>
+    <ClInclude Include="..\Siv3D\include\Siv3D\Formatter.hpp">
+      <Filter>include\Siv3D</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="..\Siv3D\src\Siv3D-Platform\WindowsDesktop\Siv3DMain.cpp">
@@ -388,9 +391,6 @@
     </ClCompile>
     <ClCompile Include="..\Siv3D\src\Siv3D\StringView\SivStringView.cpp">
       <Filter>src\Siv3D\StringView</Filter>
-    </ClCompile>
-    <ClCompile Include="..\Siv3D\src\Siv3D\Format\SivFormat.cpp">
-      <Filter>src\Siv3D\Format</Filter>
     </ClCompile>
     <ClCompile Include="..\Siv3D\src\Siv3D\IntFormatter\SivIntFormatter.cpp">
       <Filter>src\Siv3D\IntFormatter</Filter>
@@ -466,6 +466,9 @@
     </ClCompile>
     <ClCompile Include="..\Siv3D\src\Siv3D-Platform\WindowsDesktop\DLL\SivDLL.cpp">
       <Filter>src\Siv3D-Platform\WindowsDesktop\DLL</Filter>
+    </ClCompile>
+    <ClCompile Include="..\Siv3D\src\Siv3D\Formatter\SivFormatter.cpp">
+      <Filter>src\Siv3D\Formatter</Filter>
     </ClCompile>
   </ItemGroup>
 </Project>
